refactor(test): extract non-test methods from AbstractJpaMethodTest
