+++ components/camel-jpa/src/test/java/org/apache/camel/component/jpa/AbstractJpaMethodTest.java
@@ -23,39 +23,21 @@ import java.util.concurrent.TimeUnit;
 
 import jakarta.persistence.EntityManager;
 
-import org.apache.camel.Consumer;
 import org.apache.camel.Exchange;
 import org.apache.camel.Processor;
 import org.apache.camel.examples.Address;
 import org.apache.camel.examples.Customer;
-import org.apache.camel.test.junit5.CamelTestSupport;
-import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Test;
-import org.springframework.transaction.TransactionStatus;
-import org.springframework.transaction.support.TransactionCallback;
-import org.springframework.transaction.support.TransactionTemplate;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-public abstract class AbstractJpaMethodTest extends CamelTestSupport {
+public abstract class AbstractJpaMethodTest extends AbstractJpaMethodSupport {
 
-    protected JpaEndpoint endpoint;
-    protected EntityManager entityManager;
-    protected TransactionTemplate transactionTemplate;
-    protected Consumer consumer;
     protected Customer receivedCustomer;
 
     abstract boolean usePersist();
-
-    @Override
-    @AfterEach
-    public void tearDown() {
-        if (entityManager != null) {
-            entityManager.close();
-        }
-    }
 
     @Test
     public void produceNewEntity() throws Exception {
@@ -149,47 +131,4 @@ public abstract class AbstractJpaMethodTest extends CamelTestSupport {
         assertEntitiesInDatabase(0, Address.class.getName());
     }
 
-    protected void setUp(String endpointUri) throws Exception {
-        endpoint = context.getEndpoint(endpointUri, JpaEndpoint.class);
-
-        transactionTemplate = endpoint.createTransactionTemplate();
-        entityManager = endpoint.getEntityManagerFactory().createEntityManager();
-
-        transactionTemplate.execute(new TransactionCallback<Object>() {
-            public Object doInTransaction(TransactionStatus status) {
-                entityManager.joinTransaction();
-                entityManager.createQuery("delete from " + Customer.class.getName()).executeUpdate();
-                return null;
-            }
-        });
-
-        assertEntitiesInDatabase(0, Customer.class.getName());
-        assertEntitiesInDatabase(0, Address.class.getName());
-    }
-
-    protected void save(final Object persistable) {
-        transactionTemplate.execute(new TransactionCallback<Object>() {
-            public Object doInTransaction(TransactionStatus status) {
-                entityManager.joinTransaction();
-                entityManager.persist(persistable);
-                entityManager.flush();
-                return null;
-            }
-        });
-    }
-
-    protected void assertEntitiesInDatabase(int count, String entity) {
-        List<?> results = entityManager.createQuery("select o from " + entity + " o").getResultList();
-        assertEquals(count, results.size());
-    }
-
-    protected Customer createDefaultCustomer() {
-        Customer customer = new Customer();
-        customer.setName("Christian Mueller");
-        Address address = new Address();
-        address.setAddressLine1("Hahnstr. 1");
-        address.setAddressLine2("60313 Frankfurt am Main");
-        customer.setAddress(address);
-        return customer;
-    }
 }
